fix: correct LibRaw parameter implementations based on official documentation
## Core LibRaw Parameter Fixes:

**Exposure Correction:**
- Fix exp_shift calculation: now uses proper 2^exposure formula
- Correct range: -3.0 to +3.0 stops (was -4.0 to +4.0)
- Add exp_preser parameter for highlight preservation

**Gamma Correction:**
- Fix gamm array settings: [0] = 1/gamma, [1] = toe slope (4.5)
- Implement BT.709 standard gamma curve properly

**Color Temperature & Tint:**
- Implement proper Planckian locus-based Kelvin to RGB conversion
- Add scientific color temperature calculation with blackbody radiation
- Support tint adjustment through green channel multiplication
- Only apply manual WB when both camera WB and auto WB are disabled

**Highlight Recovery:**
- Change to Int32 type to match LibRaw highlight modes (0-9)
- 0: clip, 1: unclip, 2: blend, 3-9: rebuild algorithms
- Add IntSliderControl component for integer parameter UI

**Additional LibRaw Parameters:**
- Add useAutoWB for automatic white balance
- Add noAutoBright to disable automatic brightness
- Add fourColorRGB for separate green channel interpolation
- Proper output format settings (sRGB, 8-bit)

## UI Improvements:
- Add IntSliderControl for integer-based parameters
- Update highlight recovery range to 0-9 modes
- Maintain existing UI responsiveness and background processing

+++ RawSwift-example/RawSwift-example/RawImageViewModel.swift
@@ -47,7 +47,7 @@ class RawImageViewModel: ObservableObject {
     @Published var gamma: Float = 2.2 {
         didSet { updateProcessingParams() }
     }
-    @Published var highlightRecovery: Float = 0.0 {
+    @Published var highlightRecovery: Int32 = 0 {
         didSet { updateProcessingParams() }
     }
     @Published var shadowRecovery: Float = 0.0 {
@@ -193,7 +193,7 @@ class RawImageViewModel: ObservableObject {
         contrast = 1.0
         saturation = 1.0
         gamma = 2.2
-        highlightRecovery = 0.0
+        highlightRecovery = 0
         shadowRecovery = 0.0
         useCameraWB = true
         demosaicAlgorithm = .dcbInterpolation
